+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # dropwizard-json-exceptions
 
+## 1.1.0 Release Date TBD
+
+* Added configurable MediaType return types for the 2 ExceptionMappers (they're still defaulting to JSON)
+* Added ErrorMessage JSON to all types of JsonProcessingExceptionMapper responses
+
 ## 1.0.0 2015/11/18
 
 * Initial copy of the JsonProcessingExceptionMapper from https://goo.gl/uhbxtQ